Доработка функционала для детального просмотра фильма
