+++ src/view/abstract-view.js
@@ -22,7 +22,10 @@ class AbstractView {
   }
 
   removeElement = () => {
-    this.#element.remove();
+    if (this.#element) {
+      this.#element.remove();
+      this.#element = null;
+    }
   }
 }
