+++ src/view/movie-details-form-view.js
@@ -1,0 +1,11 @@
+import AbstractEventView from './abstract-event-view';
+
+const getFormTemplate = () => '<form class="film-details__inner" action="" method="get"></form>';
+
+class MovieDetailsFormView extends AbstractEventView {
+  get template() {
+    return getFormTemplate();
+  }
+}
+
+export default MovieDetailsFormView;
